refactor: move schematic from common to express-engine
Previously, this schematic used to be shared and hence why it was in common. Now this is only used in express-engine.
