+++ modules/common/BUILD.bazel
@@ -23,7 +23,6 @@ ng_package(
     srcs = [":package.json"],
     entry_point = ":index.ts",
     nested_packages = [
-        "//modules/common/schematics:npm_package",
         "//modules/common/clover/server:npm_package",
     ],
     readme_md = ":README.md",
