--- modules/common/schematics/BUILD.bazel
@@ -1,64 +1,0 @@
-load("//tools:defaults.bzl", "jasmine_node_test", "ng_test_library", "pkg_npm", "ts_library")
-
-package(default_visibility = ["//visibility:public"])
-
-filegroup(
-    name = "schematics_assets",
-    srcs = glob([
-        "**/files/**/*",
-        "**/*.json",
-    ]),
-)
-
-ts_library(
-    name = "schematics",
-    srcs = glob(
-        ["**/*.ts"],
-        exclude = [
-            "testing/**/*",
-            "**/*.spec.ts",
-            "**/files/**/*",
-        ],
-    ),
-    module_name = "@nguniversal/common/schematics",
-    tsconfig = ":tsconfig.json",
-    deps = [
-        "@npm//@angular-devkit/core",
-        "@npm//@angular-devkit/schematics",
-        "@npm//@schematics/angular",
-        "@npm//rxjs",
-        "@npm//typescript",
-    ],
-)
-
-# This package is intended to be combined into the main @nguniversal/common package as a dep.
-pkg_npm(
-    name = "npm_package",
-    srcs = [":schematics_assets"],
-    deps = [":schematics"],
-)
-
-### Testing rules
-jasmine_node_test(
-    name = "unit_tests",
-    srcs = [":schematics_test_sources"],
-    data = [":schematics_assets"],
-)
-
-ng_test_library(
-    name = "schematics_test_sources",
-    srcs = glob(
-        [
-            "testing/*.ts",
-            "**/*.spec.ts",
-        ],
-        exclude = ["**/files/**/*"],
-    ),
-    tsconfig = ":tsconfig.json",
-    deps = [
-        ":schematics",
-        "@npm//@angular-devkit/core",
-        "@npm//@angular-devkit/schematics",
-        "@npm//@schematics/angular",
-    ],
-)
