--- modules/common/schematics/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "./",
-    "lib": ["es2019", "dom"],
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "outDir": "../../../dist/schematics",
-    "noEmitOnError": false,
-    "skipDefaultLibCheck": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "target": "ES2019",
-    "types": ["jasmine", "node"],
-    "paths": {
-      "@nguniversal/*": ["../../*"]
-    }
-  },
-  "exclude": ["**/*.spec.ts", "*/files/**/*"],
-  "bazelOptions": {
-    "suppressTsconfigOverrideWarnings": true
-  }
-}
