+++ modules/express-engine/schematics/BUILD.bazel
@@ -25,11 +25,9 @@ ts_library(
     module_name = "@nguniversal/express-engine/schematics",
     tsconfig = ":tsconfig.json",
     deps = [
-        "//modules/common/schematics",
         "@npm//@angular-devkit/core",
         "@npm//@angular-devkit/schematics",
         "@npm//@schematics/angular",
-        "@npm//@types/jasmine",
         "@npm//rxjs",
         "@npm//typescript",
     ],
@@ -49,7 +47,6 @@ jasmine_node_test(
     srcs = [":schematics_test_sources"],
     data = [
         ":schematics_assets",
-        "//modules/common/schematics:npm_package",
     ],
 )
 
@@ -65,7 +62,6 @@ ng_test_library(
     tsconfig = ":tsconfig.json",
     deps = [
         ":schematics",
-        "//modules/common/schematics",
         "@npm//@angular-devkit/core",
         "@npm//@angular-devkit/schematics",
         "@npm//@schematics/angular",
