+++ modules/express-engine/schematics/install/schema.ts
@@ -28,10 +28,6 @@ export interface Schema {
    */
   serverPort?: number;
   /**
-   * The name of the TypeScript configuration file.
-   */
-  tsconfigFileName?: string;
-  /**
    * The name of the application directory.
    */
   appDir?: string;
